+++ docs/src/release_notes.md
@@ -9,15 +9,19 @@ All changes are grouped by type and the latest version appears first.
 ### Added
 * feat(upload): add upload description
 * feat(upload): request new uploads
+* feat(upload): edit files
 * feat(homework/event): add homework/event pinning
 
 ### Fixed
 * fix(rate-limit): increase proxy and upadte nginx.config
 * fix(subject): subject data not stringified correctly
+* fix(account): send 201 instead of 200 at /register
 
 ### Changed
+* chore(getdata): event, homework and upload metadata order by more values
 * chore(docs): migrate to mdbook and github actions/pages for deployment
 * chore(ci): update CI build test
+* chore(package): update package.json version to v2.2.5
 
 ---